requirements.txt
accumulation-tree==0.6
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
docutils==0.15.2
hurry.filesize==0.9
idna==2.8
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.0
numpy==1.16.4
pandas==0.24.2
pkginfo==1.5.0.1
Pygments==2.4.2
Pympler==0.7
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
pyudorandom==1.0.0
PyYAML==5.1.2
pyyml==0.0.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
subprocess32==3.5.4
tdigest==0.5.2.2
tqdm==4.36.1
twine==1.13.0
urllib3==1.25.6
webencodings==0.5.1
